+++ internal/dhcpd/db.go
@@ -128,7 +128,7 @@ func normalizeLeases(staticLeases, dynLeases []*Lease) []*Lease {
 func (s *Server) dbStore() {
 	var leases []leaseJSON
 
-	leases4 := s.srv4.GetLeasesRef()
+	leases4 := s.srv4.getLeasesRef()
 	for _, l := range leases4 {
 		if l.Expiry.Unix() == 0 {
 			continue
@@ -143,7 +143,7 @@ func (s *Server) dbStore() {
 	}
 
 	if s.srv6 != nil {
-		leases6 := s.srv6.GetLeasesRef()
+		leases6 := s.srv6.getLeasesRef()
 		for _, l := range leases6 {
 			if l.Expiry.Unix() == 0 {
 				continue